ExpectCovar-class {coin}	R Documentation

Class "ExpectCovar"

Description

Objects of class "ExpectCovar" represent the expectation and covariance of the linear statistic.

Objects from the Class

Objects can be created by calls of the form

     new("ExpectCovar", \dots).

Slots

expectation:

Object of class "numeric". The expectation of the linear statistic.

covariance:

Object of class "matrix". The covariance matrix of the linear statistic.

dimension:

Object of class "integer". The dimension of the linear statistic.

Known Subclasses

Class "ExpectCovarInfluence", directly.

Methods

initialize

signature(.Object = "ExpectCovar"): See the documentation for initialize (in package methods) for details.

Note

Normally, this class is used internally only.

[Package coin version 1.1-2 Index]