ExpectCovarInfluence-class {coin}	R Documentation

Class "ExpectCovarInfluence"

Description

Objects of class "ExpectCovarInfluence" represent the expectation and covariance of the influence function.

Objects from the Class

Objects can be created by calls of the form

     new("ExpectCovarInfluence", \dots).

Slots

sumweights:

Object of class "numeric". The sum of the case weights.

expectation:

Object of class "numeric". The expectation of the influence function.

covariance:

Object of class "matrix". The covariance matrix of the influence function.

dimension:

Object of class "integer". The dimension of the influence function.

Extends

Class "ExpectCovar", directly.

Methods

initialize

signature(.Object = "ExpectCovarInfluence"): See the documentation for initialize (in package methods) for details.

Note

Normally, this class is used internally only.

[Package coin version 1.1-2 Index]