IndependenceLinearStatistic-class {coin}	R Documentation

Class "IndependenceLinearStatistic"

Description

Objects of class "IndependenceLinearStatistic" represent the linear statistic and the transformed and original data structures corresponding to an independence problem.

Objects from the Class

Objects can be created by calls of the form

     new("IndependenceLinearStatistic", object, \dots)

where object is an object of class "IndependenceTestProblem".

Slots

linearstatistic:

Object of class "matrix". The linear statistic for each block.

expectation:

Object of class "matrix". The expectation of the linear statistic for each block.

covariance:

Object of class "matrix". The lower triangular elements of the covariance of the linear statistic for each block.

xtrans:

Object of class "matrix". The transformed x.

ytrans:

Object of class "matrix". The transformed y.

xtrafo:

Object of class "function". The regression function for x.

ytrafo:

Object of class "function". The influence function for y.

x:

Object of class "data.frame". The variables x.

y:

Object of class "data.frame". The variables y.

block:

Object of class "factor". The block structure.

weights:

Object of class "numeric". The case weights.

Extends

Class "IndependenceTestProblem", directly.
Class "IndependenceProblem", by class "IndependenceTestProblem", distance 2.

Known Subclasses

Class "IndependenceTestStatistic", directly.
Class "MaxTypeIndependenceTestStatistic", by class "IndependenceTestStatistic", distance 2.
Class "QuadTypeIndependenceTestStatistic", by class "IndependenceTestStatistic", distance 2.
Class "ScalarIndependenceTestStatistic", by class "IndependenceTestStatistic", distance 2.

Methods

covariance

signature(object = "IndependenceLinearStatistic"): See the documentation for covariance for details.

expectation

signature(object = "IndependenceLinearStatistic"): See the documentation for expectation for details.

initialize

signature(.Object = "IndependenceLinearStatistic"): See the documentation for initialize (in package methods) for details.

statistic

signature(object = "IndependenceLinearStatistic"): See the documentation for statistic for details.

variance

signature(object = "IndependenceLinearStatistic"): See the documentation for variance for details.

[Package coin version 1.4-1 Index]