IndependenceProblem-class {coin}	R Documentation

Class "IndependenceProblem"

Description

Objects of class "IndependenceProblem" represent the data structure corresponding to an independence problem.

Objects from the Class

Objects can be created by calls of the form

     new("IndependenceProblem", x, y, block = NULL, weights = NULL, \dots)

where x and y are data frames containing the variables X and Y respectively, block is an optional factor representing the block structure b and weights is an optional integer vector corresponding to the case weights w.

Slots

x:

Object of class "data.frame". The variables x.

y:

Object of class "data.frame". The variables y.

block:

Object of class "factor". The block structure.

weights:

Object of class "numeric". The case weights.

Known Subclasses

Class "IndependenceTestProblem", directly.
Class "SymmetryProblem", directly.
Class "IndependenceLinearStatistic", by class "IndependenceTestProblem", distance 2.
Class "IndependenceTestStatistic", by class "IndependenceTestProblem", distance 3.
Class "MaxTypeIndependenceTestStatistic", by class "IndependenceTestProblem", distance 4.
Class "QuadTypeIndependenceTestStatistic", by class "IndependenceTestProblem", distance 4.
Class "ScalarIndependenceTestStatistic", by class "IndependenceTestProblem", distance 4.

Methods

initialize

signature(.Object = "IndependenceProblem"): See the documentation for initialize (in package methods) for details.

[Package coin version 1.4-1 Index]