IndependenceTestProblem-class {coin}	R Documentation

Class "IndependenceTestProblem"

Description

Objects of class "IndependenceTestProblem" represent the transformed and original data structures corresponding to an independence problem.

Objects from the Class

Objects can be created by calls of the form

     new("IndependenceTestProblem", object, xtrafo = trafo, ytrafo = trafo, \dots)

where object is an object of class "IndependenceProblem", xtrafo is the regression function g(X) and ytrafo is the influence function h(Y).

Slots

xtrans:

Object of class "matrix". The transformed x.

ytrans:

Object of class "matrix". The transformed y.

xtrafo:

Object of class "function". The regression function for x.

ytrafo:

Object of class "function". The influence function for y.

x:

Object of class "data.frame". The variables x.

y:

Object of class "data.frame". The variables y.

block:

Object of class "factor". The block structure.

weights:

Object of class "numeric". The case weights.

Extends

Class "IndependenceProblem", directly.

Known Subclasses

Class "IndependenceLinearStatistic", directly.
Class "IndependenceTestStatistic", by class "IndependenceLinearStatistic", distance 2.
Class "MaxTypeIndependenceTestStatistic", by class "IndependenceTestStatistic", distance 3.
Class "QuadTypeIndependenceTestStatistic", by class "IndependenceTestStatistic", distance 3.
Class "ScalarIndependenceTestStatistic", by class "IndependenceTestStatistic", distance 3.

Methods

initialize

signature(.Object = "IndependenceTestProblem"): See the documentation for initialize (in package methods) for details.

[Package coin version 1.4-1 Index]