IndependenceTestStatistic-class {coin}	R Documentation

Class "IndependenceTestStatistic" and its subclasses

Description

Objects of class "IndependenceTestStatistic" and its subclasses "MaxTypeIndependenceTestStatistic", "QuadTypeIndependenceTestStatistic" and "ScalarIndependenceTestStatistic" represent the test statistic, the linear statistic, and the transformed and original data structures corresponding to an independence problem.

Objects from the Class

Class "IndependenceTestStatistic" is a virtual class, so objects cannot be created from it directly.

Objects can be created by calls of the form

     new("MaxTypeIndependenceTestStatistic", object,
         alternative = c("two.sided", "less", "greater"), \dots),

     new("QuadTypeIndependenceTestStatistic", object, paired = FALSE, \dots)

and

     new("ScalarIndependenceTestStatistic", object,
         alternative = c("two.sided", "less", "greater"), paired = FALSE, \dots)

where object is an object of class "IndependenceLinearStatistic", alternative is a character specifying the direction of the alternative hypothesis and paired is a logical indicating that paired data have been transformed in such a way that the (unstandardized) linear statistic is the sum of the absolute values of the positive differences between the paired observations.

Slots

For objects of classes "IndependenceTestStatistic", "MaxTypeIndependenceTestStatistic", "QuadTypeIndependenceTestStatistic" or "ScalarIndependenceTestStatistic":

teststatistic:

Object of class "numeric". The test statistic.

standardizedlinearstatistic:

Object of class "numeric". The standardized linear statistic.

linearstatistic:

Object of class "matrix". The linear statistic for each block.

expectation:

Object of class "matrix". The expectation of the linear statistic for each block.

covariance:

Object of class "matrix". The lower triangular elements of the covariance of the linear statistic for each block.

xtrans:

Object of class "matrix". The transformed x.

ytrans:

Object of class "matrix". The transformed y.

xtrafo:

Object of class "function". The regression function for x.

ytrafo:

Object of class "function". The influence function for y.

x:

Object of class "data.frame". The variables x.

y:

Object of class "data.frame". The variables y.

block:

Object of class "factor". The block structure.

weights:

Object of class "numeric". The case weights.

Additionally, for objects of classes "MaxTypeIndependenceTest" or "ScalarIndependenceTest":

alternative:

Object of class "character". The direction of the alternative hypothesis.

Additionally, for objects of class "QuadTypeIndependenceTest":

covarianceplus:

Object of class "numeric". The lower triangular elements of the Moore-Penrose inverse of the covariance of the linear statistic.

df:

Object of class "numeric". The rank of the covariance matrix.

Additionally, for objects of classes "QuadTypeIndependenceTest" or "ScalarIndependenceTest":

paired:

Object of class "logical". The indicator for paired test statistics.

Extends

For objects of class "IndependenceTestStatistic":
Class "IndependenceLinearStatistic", directly.
Class "IndependenceTestProblem", by class "IndependenceLinearStatistic", distance 2.
Class "IndependenceProblem", by class "IndependenceLinearStatistic", distance 3.

For objects of classes "MaxTypeIndependenceTestStatistic", "QuadTypeIndependenceTestStatistic" or "ScalarIndependenceTestStatistic":
Class "IndependenceTestStatistic", directly.
Class "IndependenceLinearStatistic", by class "IndependenceTestStatistic", distance 2.
Class "IndependenceTestProblem", by class "IndependenceTestStatistic", distance 3.
Class "IndependenceProblem", by class "IndependenceTestStatistic", distance 4.

Known Subclasses

For objects of class "IndependenceTestStatistic":
Class "MaxTypeIndependenceTestStatistic", directly.
Class "QuadTypeIndependenceTestStatistic", directly.
Class "ScalarIndependenceTestStatistic", directly.

Methods

ApproxNullDistribution

signature(object = "MaxTypeIndependenceTestStatistic"): See the documentation for ApproxNullDistribution for details.

ApproxNullDistribution

signature(object = "QuadTypeIndependenceTestStatistic"): See the documentation for ApproxNullDistribution for details.

ApproxNullDistribution

signature(object = "ScalarIndependenceTestStatistic"): See the documentation for ApproxNullDistribution for details.

AsymptNullDistribution

signature(object = "MaxTypeIndependenceTestStatistic"): See the documentation for AsymptNullDistribution for details.

AsymptNullDistribution

signature(object = "QuadTypeIndependenceTestStatistic"): See the documentation for AsymptNullDistribution for details.

AsymptNullDistribution

signature(object = "ScalarIndependenceTestStatistic"): See the documentation for AsymptNullDistribution for details.

ExactNullDistribution

signature(object = "QuadTypeIndependenceTestStatistic"): See the documentation for ExactNullDistribution for details.

ExactNullDistribution

signature(object = "ScalarIndependenceTestStatistic"): See the documentation for ExactNullDistribution for details.

covariance

signature(object = "QuadTypeIndependenceTestStatistic"): See the documentation for covariance for details.

initialize

signature(.Object = "IndependenceTestStatistic"): See the documentation for initialize (in package methods) for details.

initialize

signature(.Object = "MaxTypeIndependenceTestStatistic"): See the documentation for initialize (in package methods) for details.

initialize

signature(.Object = "QuadTypeIndependenceTestStatistic"): See the documentation for initialize (in package methods) for details.

initialize

signature(.Object = "ScalarIndependenceTestStatistic"): See the documentation for initialize (in package methods) for details.

statistic

signature(object = "IndependenceTestStatistic"): See the documentation for statistic for details.

[Package coin version 1.4-1 Index]