NullDistribution-methods {coin}	R Documentation

Computation of the Reference Distribution

Description

Methods for computation of the asymptotic, approximative (Monte Carlo) and exact reference distribution.

Usage

## S4 method for signature 'MaxTypeIndependenceTestStatistic'
AsymptNullDistribution(object, ...)
## S4 method for signature 'QuadTypeIndependenceTestStatistic'
AsymptNullDistribution(object, ...)
## S4 method for signature 'ScalarIndependenceTestStatistic'
AsymptNullDistribution(object, ...)

## S4 method for signature 'MaxTypeIndependenceTestStatistic'
ApproxNullDistribution(object, nresample = 10000L, B, ...)
## S4 method for signature 'QuadTypeIndependenceTestStatistic'
ApproxNullDistribution(object, nresample = 10000L, B, ...)
## S4 method for signature 'ScalarIndependenceTestStatistic'
ApproxNullDistribution(object, nresample = 10000L, B, ...)

## S4 method for signature 'QuadTypeIndependenceTestStatistic'
ExactNullDistribution(object, algorithm = c("auto", "shift", "split-up"), ...)
## S4 method for signature 'ScalarIndependenceTestStatistic'
ExactNullDistribution(object, algorithm = c("auto", "shift", "split-up"), ...)

Arguments

object
an object from which the asymptotic, approximative (Monte Carlo) or exact reference distribution can be computed.

nresample
a positive integer, the number of Monte Carlo replicates used for the computation of the approximative reference distribution. Defaults to 10000L.

B
deprecated, use nresample instead.

algorithm
a character, the algorithm used for the computation of the exact reference distribution: either "auto" (default), "shift" or "split-up".

...
further arguments to be passed to or from methods.

Details

The methods AsymptNullDistribution, ApproxNullDistribution and ExactNullDistribution compute the asymptotic, approximative (Monte Carlo) and exact reference distribution respectively.

Value

An object of class "AsymptNullDistribution", "ApproxNullDistribution" or "ExactNullDistribution".

Note

In versions of coin prior to 1.3-0, the number of Monte Carlo replicates in ApproxNullDistribution() was specified using the now deprecated B argument. This will be made defunct and removed in a future release. It has been replaced by the nresample argument (for conformity with the libcoin, party and partykit packages).

[Package coin version 1.4-1 Index]