PValue-class {coin}	R Documentation

Class "PValue"

Description

Objects of class "PValue" represent the p-value, mid-p-value and p-value interval of the reference distribution.

Objects from the Class

Objects can be created by calls of the form

     new("PValue", \dots).

Slots

name:

Object of class "character". The name of the reference distribution.

p:

Object of class "function". The distribution function of the reference distribution.

pvalue:

Object of class "function". The p-value function of the reference distribution.

Methods

pvalue

signature(object = "PValue"): See the documentation for pvalue for details.

Note

Starting with coin version 1.3-0, this class is deprecated and will be replaced by class "NullDistribution". It will be made defunct and removed in a future release.

[Package coin version 1.4-1 Index]