VarCovar-class {coin}	R Documentation

Class "VarCovar" and its subclasses

Description

Objects of class "VarCovar" and its subclasses "CovarianceMatrix" and "Variance" represent the covariance and variance, respectively, of the linear statistic.

Objects from the Class

Class "VarCovar" is a virtual class defined as the class union of "CovarianceMatrix" and "Variance", so objects cannot be created from it directly.

Objects can be created by calls of the form

     new("CovarianceMatrix", covariance, \dots)

and

     new("Variance", variance, \dots)

where covariance is a covariance matrix and variance is numeric vector containing the diagonal elements of the covariance matrix.

Slots

For objects of class "CovarianceMatrix":

covariance:

Object of class "matrix". The covariance matrix.

For objects of class "Variance":

variance:

Object of class "numeric". The diagonal elements of the covariance matrix.

Extends

For objects of classes "CovarianceMatrix" or "Variance":
Class "VarCovar", directly.

Known Subclasses

For objects of class "VarCovar":
Class "CovarianceMatrix", directly.
Class "Variance", directly.

Methods

covariance

signature(object = "CovarianceMatrix"): See the documentation for covariance for details.

initialize

signature(.Object = "CovarianceMatrix"): See the documentation for initialize (in package methods) for details.

initialize

signature(.Object = "Variance"): See the documentation for initialize (in package methods) for details.

variance

signature(object = "CovarianceMatrix"): See the documentation for variance for details.

variance

signature(object = "Variance"): See the documentation for variance for details.

Note

Starting with coin version 1.4-0, this class is deprecated. It will be made defunct and removed in a future release.

[Package coin version 1.4-1 Index]