alzheimer {coin}	R Documentation

Smoking and Alzheimer's Disease

Description

A case-control study of smoking and Alzheimer's disease.

Usage

alzheimer

Format

A data frame with 538 observations on 3 variables.

smoking

a factor with levels "None", "<10", "10-20" and ">20" (cigarettes per day).

disease

a factor with levels "Alzheimer", "Other dementias" and "Other diagnoses".

gender

a factor with levels "Female" and "Male".

Details

Subjects with Alzheimer's disease are compared to two different control groups with respect to smoking history. The data are given in Salib and Hillier (1997, Tab. 4).

Source

Salib, E. and Hillier, V. (1997). A case-control study of smoking and Alzheimer's disease. International Journal of Geriatric Psychiatry 12(3), 295–300. doi: 10.1002/(SICI)1099-1166(199703)12:3<295::AID-GPS476>3.0.CO;2-3

References

Hothorn, T., Hornik, K., van de Wiel, M. A. and Zeileis, A. (2006). A Lego system for conditional inference. The American Statistician 60(3), 257–263. doi: 10.1198/000313006X118430

Examples

## Spineplots
op <- par(no.readonly = TRUE) # save current settings
layout(matrix(1:2, ncol = 2))
spineplot(disease ~ smoking, data = alzheimer,
          subset = gender == "Male", main = "Male")
spineplot(disease ~ smoking, data = alzheimer,
          subset = gender == "Female", main = "Female")
par(op) # reset

## Asymptotic Cochran-Mantel-Haenszel test
cmh_test(disease ~ smoking | gender, data = alzheimer)

[Package coin version 1.4-1 Index]