hohnloser {coin}	R Documentation

Left Ventricular Ejection Fraction

Description

Left ventricular ejection fraction in patients with malignant ventricular tachyarrhythmias including recurrence-free month and censoring.

Usage

hohnloser

Format

A data frame with 94 observations on 3 variables.

EF

ejection fraction (%).

time

recurrence-free month.

event

status indicator for time: 0 for right-censored observations and 1 otherwise.

Details

The data was used by Lausen and Schumacher (1992) to illustrate the use of maximally selected statistics.

Source

Hohnloser, S. H., Raeder, E. A., Podrid, P. J., Graboys, T. B. and Lown, B. (1987). Predictors of antiarrhythmic drug efficacy in patients with malignant ventricular tachyarrhythmias. American Heart Journal 114(1 Pt 1), 1–7. doi: 10.1016/0002-8703(87)90299-7

References

Lausen, B. and Schumacher, M. (1992). Maximally selected rank statistics. Biometrics 48(1), 73–85. doi: 10.2307/2532740

Examples

## Asymptotic maximally selected logrank statistics
maxstat_test(Surv(time, event) ~ EF, data = hohnloser)

[Package coin version 1.4-1 Index]