neuropathy {coin}	R Documentation

Acute Painful Diabetic Neuropathy

Description

The logarithm of the ratio of pain scores measured at baseline and after four weeks in a control group and a treatment group.

Usage

neuropathy

Format

A data frame with 58 observations on 2 variables.

pain

pain scores: ln(baseline / final).

group

a factor with levels "control" and "treat".

Details

Data from Conover and Salsburg (1988, Tab. 1).

Source

Conover, W. J. and Salsburg, D. S. (1988). Locally most powerful tests for detecting treatment effects when only a subset of patients can be expected to “respond” to treatment. Biometrics 44(1), 189–196. doi: 10.2307/2531906

Examples

## Conover and Salsburg (1988, Tab. 2)

## One-sided approximative Fisher-Pitman test
oneway_test(pain ~ group, data = neuropathy,
            alternative = "less",
            distribution = approximate(nresample = 10000))

## One-sided approximative Wilcoxon-Mann-Whitney test
wilcox_test(pain ~ group, data = neuropathy,
            alternative = "less",
            distribution = approximate(nresample = 10000))

## One-sided approximative Conover-Salsburg test
oneway_test(pain ~ group, data = neuropathy,
            alternative = "less",
            distribution = approximate(nresample = 10000),
            ytrafo = function(data)
                trafo(data, numeric_trafo = consal_trafo))

## One-sided approximative maximum test for a range of 'a' values
it <- independence_test(pain ~ group, data = neuropathy,
                        alternative = "less",
                        distribution = approximate(nresample = 10000),
                        ytrafo = function(data)
                            trafo(data, numeric_trafo = function(y)
                                consal_trafo(y, a = 2:7)))
pvalue(it, method = "single-step")

[Package coin version 1.4-1 Index]