photocar {coin}	R Documentation

Multiple Dosing Photococarcinogenicity Experiment

Description

Survival time, time to first tumor, and total number of tumors in three groups of animals in a photococarcinogenicity study.

Usage

photocar

Format

A data frame with 108 observations on 6 variables.

group

a factor with levels "A", "B", and "C".

ntumor

total number of tumors.

time

survival time.

event

status indicator for time: FALSE for right-censored observations and TRUE otherwise.

dmin

time to first tumor.

tumor

status indicator for dmin: FALSE when no tumor was observed and TRUE otherwise.

Details

The animals were exposed to different levels of ultraviolet radiation (UVR) exposure (group A: topical vehicle and 600 Robertson–Berger units of UVR, group B: no topical vehicle and 600 Robertson–Berger units of UVR and group C: no topical vehicle and 1200 Robertson–Berger units of UVR). The data are taken from Tables 1 to 3 in Molefe et al. (2005).

The main interest is testing the global null hypothesis of no treatment effect with respect to survival time, time to first tumor and number of tumors. (Molefe et al., 2005, also analysed the detection time of tumors, but that data is not given here.) In case the global null hypothesis can be rejected, the deviations from the partial null hypotheses are of special interest.

Source

Molefe, D. F., Chen, J. J., Howard, P. C., Miller, B. J., Sambuco, C. P., Forbes, P. D. and Kodell, R. L. (2005). Tests for effects on tumor frequency and latency in multiple dosing photococarcinogenicity experiments. Journal of Statistical Planning and Inference 129(1–2), 39–58. doi: 10.1016/j.jspi.2004.06.038

References

Hothorn, T., Hornik, K., van de Wiel, M. A. and Zeileis, A. (2006). A Lego system for conditional inference. The American Statistician 60(3), 257–263. doi: 10.1198/000313006X118430

Examples

## Plotting data
op <- par(no.readonly = TRUE) # save current settings
layout(matrix(1:3, ncol = 3))
with(photocar, {
    plot(survfit(Surv(time, event) ~ group),
         lty =  1:3, xmax = 50, main = "Survival Time")
    legend("bottomleft", lty = 1:3, levels(group), bty = "n")
    plot(survfit(Surv(dmin, tumor) ~ group),
         lty = 1:3, xmax = 50, main = "Time to First Tumor")
    legend("bottomleft", lty = 1:3, levels(group), bty = "n")
    boxplot(ntumor ~ group, main = "Number of Tumors")
})
par(op) # reset

## Approximative multivariate (all three responses) test
it <- independence_test(Surv(time, event) + Surv(dmin, tumor) + ntumor ~ group,
                        data = photocar,
                        distribution = approximate(nresample = 10000))

## Global p-value
pvalue(it)

## Why was the global null hypothesis rejected?
statistic(it, type = "standardized")
pvalue(it, method = "single-step")

[Package coin version 1.4-1 Index]