statistic-methods {coin}	R Documentation

Extraction of the Test Statistic and the Linear Statistic

Description

Methods for extraction of the test statistic and the linear statistic.

Usage

## S4 method for signature 'IndependenceLinearStatistic'
statistic(object, type = c("test", "linear", "centered", "standardized"),
          partial = FALSE, ...)
## S4 method for signature 'IndependenceTestStatistic'
statistic(object, type = c("test", "linear", "centered", "standardized"),
          partial = FALSE, ...)
## S4 method for signature 'IndependenceTest'
statistic(object, type = c("test", "linear", "centered", "standardized"),
          partial = FALSE, ...)

Arguments

object
an object from which the test statistic or the linear statistic can be extracted.

type
a character string indicating the type of statistic: either "test" (default) for the test statistic, "linear" for the unstandardized linear statistic, "centered" for the centered linear statistic or "standardized" for the standardized linear statistic.

partial
a logical indicating that the partial linear statistic for each block should be extracted. Defaults to FALSE.

...
further arguments (currently ignored).

Details

The method statistic extracts the univariate test statistic or the, possibly multivariate, linear statistic in its unstandardized, centered or standardized form.

The test statistic (type = "test") is returned by default. The unstandardized, centered or standardized linear statistic is obtained by setting type to "linear", "centered" or "standardized", respectively. For tests of conditional independence within blocks, the partial linear statistic for each block is obtained by setting partial = TRUE.

Value

The test statistic or the unstandardized, centered or standardized linear statistic extracted from object. A numeric vector, matrix or array.

Examples

## Example data
dta <- data.frame(
    y = gl(4, 5),
    x = gl(5, 4)
)

## Asymptotic Cochran-Mantel-Haenszel Test
ct <- cmh_test(y ~ x, data = dta)

## Test statistic
statistic(ct)

## The unstandardized linear statistic...
statistic(ct, type = "linear")

## ...is identical to the contingency table
xtabs(~ x + y, data = dta)

## The centered linear statistic...
statistic(ct, type = "centered")

## ...is identical to
statistic(ct, type = "linear") - expectation(ct)

## The standardized linear statistic, illustrating departures from the null
## hypothesis of independence...
statistic(ct, type = "standardized")

## ...is identical to
(statistic(ct, type = "linear") - expectation(ct)) / sqrt(variance(ct))

[Package coin version 1.4-1 Index]