GTSG {coin}	R Documentation

Gastrointestinal Tumor Study Group

Description

A randomized clinical trial in gastric cancer.

Usage

GTSG

Format

A data frame with 90 observations on 3 variables.

time

survival time (days).

event

status indicator for time: 0 for right-censored observations and 1 otherwise.

group

a factor with levels "Chemotherapy+Radiation" and "Chemotherapy".

Details

A clinical trial comparing chemotherapy alone versus a combination of chemotherapy and radiation therapy in the treatment of locally advanced, nonresectable gastric carcinoma.

Note

There is substantial separation between the estimated survival distributions at 8 to 10 months, but by month 26 the distributions intersect.

Source

Stablein, D. M., Carter, W. H., Jr. and Novak, J. W. (1981). Analysis of survival data with nonproportional hazard functions. Controlled Clinical Trials 2(2), 149–159. doi: 10.1016/0197-2456(81)90005-2

References

Moreau, T., Maccario, J., Lellouch, J. and Huber, C. (1992). Weighted log rank statistics for comparing two distributions. Biometrika 79(1), 195–198. doi: 10.1093/biomet/79.1.195

Shen, W. and Le, C. T. (2000). Linear rank tests for censored survival data. Communications in Statistics – Simulation and Computation 29(1), 21–36. doi: 10.1080/03610910008813599

Tubert-Bitter, P., Kramar, A., Chalé, J. J. and Moureau, T. (1994). Linear rank tests for comparing survival in two groups with crossing hazards. Computational Statistics & Data Analysis 18(5), 547–559. doi: 10.1016/0167-9473(94)90084-1

Examples

## Plot Kaplan-Meier estimates
plot(survfit(Surv(time / (365.25 / 12), event) ~ group, data = GTSG),
     lty = 1:2, ylab = "% Survival", xlab = "Survival Time in Months")
legend("topright", lty = 1:2,
       c("Chemotherapy+Radiation", "Chemotherapy"), bty = "n")

## Asymptotic logrank test
logrank_test(Surv(time, event) ~ group, data = GTSG)

## Asymptotic Prentice test
logrank_test(Surv(time, event) ~ group, data = GTSG, type = "Prentice")

## Asymptotic test against Weibull-type alternatives (Moreau et al., 1992)
moreau_weight <- function(time, n.risk, n.event)
    1 + log(-log(cumprod(n.risk / (n.risk + n.event))))

independence_test(Surv(time, event) ~ group, data = GTSG,
                  ytrafo = function(data)
                      trafo(data, surv_trafo = function(y)
                          logrank_trafo(y, weight = moreau_weight)))

## Asymptotic test against crossing-curve alternatives (Shen and Le, 2000)
shen_trafo <- function(x)
    ansari_trafo(logrank_trafo(x, type = "Prentice"))

independence_test(Surv(time, event) ~ group, data = GTSG,
                  ytrafo = function(data)
                      trafo(data, surv_trafo = shen_trafo))

[Package coin version 1.4-1 Index]