IndependenceTest-class {coin}	R Documentation

Class "IndependenceTest" and its subclasses

Description

Objects of class "IndependenceTest" and its subclasses "MaxTypeIndependenceTest", "QuadTypeIndependenceTest", "ScalarIndependenceTest" and "ScalarIndependenceTestConfint" represent an independence test including its original and transformed data structure, linear statistic, test statistic and reference distribution.

Objects from the Class

Objects can be created by calls of the form

    new("IndependenceTest", \dots),

    new("MaxTypeIndependenceTest", \dots),

    new("QuadTypeIndependenceTest", \dots),

    new("ScalarIndependenceTest", \dots)

and

    new("ScalarIndependenceTestConfint", \dots).

Slots

For objects of classes "IndependenceTest", "MaxTypeIndependenceTest", "QuadTypeIndependenceTest", "ScalarIndependenceTest" or "ScalarIndependenceTestConfint":

distribution:

Object of class "PValue". The reference distribution.

statistic:

Object of class "IndependenceTestStatistic". The test statistic, the linear statistic, and the transformed and original data structures.

estimates:

Object of class "list". The estimated parameters.

method:

Object of class "character". The test method.

call:

Object of class "call". The matched call.

Additionally, for objects of classes "ScalarIndependenceTest" or "ScalarIndependenceTestConfint":

parameter:

Object of class "character". The tested parameter.

nullvalue:

Object of class "numeric". The hypothesized value of the null hypothesis.

Additionally, for objects of class "ScalarIndependenceTestConfint":

confint:

Object of class "function". The confidence interval function.

conf.level:

Object of class "numeric". The confidence level.

Extends

For objects of classes "MaxTypeIndependenceTest", "QuadTypeIndependenceTest" or "ScalarIndependenceTest":
Class "IndependenceTest", directly.

For objects of class "ScalarIndependenceTestConfint":
Class "ScalarIndependenceTest", directly.
Class "IndependenceTest", by class "ScalarIndependenceTest", distance 2.

Known Subclasses

For objects of class "IndependenceTest":
Class "MaxTypeIndependenceTest", directly.
Class "QuadTypeIndependenceTest", directly.
Class "ScalarIndependenceTest", directly.
Class "ScalarIndependenceTestConfint", by class "ScalarIndependenceTest", distance 2.

For objects of class "ScalarIndependenceTest":
Class "ScalarIndependenceTestConfint", directly.

Methods

confint

signature(object = "IndependenceTest"): See the documentation for confint-methods (in package stats4) for details.

confint

signature(object = "ScalarIndependenceTestConfint"): See the documentation for confint-methods (in package stats4) for details.

covariance

signature(object = "IndependenceTest"): See the documentation for covariance for details.

dperm

signature(object = "IndependenceTest"): See the documentation for dperm for details.

expectation

signature(object = "IndependenceTest"): See the documentation for expectation for details.

midpvalue

signature(object = "IndependenceTest"): See the documentation for midpvalue for details.

pperm

signature(object = "IndependenceTest"): See the documentation for pperm for details.

pvalue

signature(object = "IndependenceTest"): See the documentation for pvalue for details.

pvalue_interval

signature(object = "IndependenceTest"): See the documentation for pvalue_interval for details.

qperm

signature(object = "IndependenceTest"): See the documentation for qperm for details.

rperm

signature(object = "IndependenceTest"): See the documentation for rperm for details.

show

signature(object = "IndependenceTest"): See the documentation for show (in package methods) for details.

show

signature(object = "MaxTypeIndependenceTest"): See the documentation for show (in package methods) for details.

show

signature(object = "QuadTypeIndependenceTest"): See the documentation for show (in package methods) for details.

show

signature(object = "ScalarIndependenceTest"): See the documentation for show (in package methods) for details.

show

signature(object = "ScalarIndependenceTestConfint"): See the documentation for show (in package methods) for details.

size

signature(object = "IndependenceTest"): See the documentation for size for details.

statistic

signature(object = "IndependenceTest"): See the documentation for statistic for details.

support

signature(object = "IndependenceTest"): See the documentation for support for details.

variance

signature(object = "IndependenceTest"): See the documentation for variance for details.

[Package coin version 1.4-1 Index]