malformations {coin}	R Documentation

Maternal Drinking and Congenital Sex Organ Malformation

Description

A subset of data from a study on the relationship between maternal alcohol consumption and congenital malformations.

Usage

malformations

Format

A data frame with 32574 observations on 2 variables.

consumption

alcohol consumption, an ordered factor with levels "0", "<1", "1-2", "3-5" and ">=6".

malformation

congenital sex organ malformation, a factor with levels "Present" and "Absent".

Details

Data from a prospective study undertaken to determine whether moderate or light drinking during the first trimester of pregnancy increases the risk for congenital malformations (Mills and Graubard, 1987). The subset given here concerns only sex organ malformation (Mills and Graubard, 1987, Tab. 4).

Graubard and Korn (1987) used this data set to illustrate that different choices of scores for ordinal variables can lead to conflicting conclusions.

Source

Mills, J. L. and Graubard, B. I. (1987). Is moderate drinking during pregnancy associated with an increased risk for malformations? Pediatrics 80(3), 309–314.

References

Graubard, B. I. and Korn, E. L. (1987). Choice of column scores for testing independence in ordered 2 x K contingency tables. Biometrics 43(2), 471–476.

Examples

## Graubard and Korn (1987, Tab. 3)

## One-sided approximative (Monte Carlo) Cochran-Armitage test
## Note: midpoint scores (p < 0.05)
midpoints <- c(0, 0.5, 1.5, 4.0, 7.0)
chisq_test(malformation ~ consumption, data = malformations,
           distribution = approximate(B = 1000), alternative = "greater",
           scores = list(consumption = midpoints))

## One-sided approximative (Monte Carlo) Cochran-Armitage test
## Note: midrank scores (p > 0.05)
midranks <- c(8557.5, 24375.5, 32013.0, 32473.0, 32555.5)
chisq_test(malformation ~ consumption, data = malformations,
           distribution = approximate(B = 1000), alternative = "greater",
           scores = list(consumption = midranks))

## One-sided approximative (Monte Carlo) Cochran-Armitage test
## Note: equally spaced scores (p > 0.05)
chisq_test(malformation ~ consumption, data = malformations,
           distribution = approximate(B = 1000), alternative = "greater")

[Package coin version 1.1-2 Index]