rotarod {coin}	R Documentation

Rotating Rats

Description

The endurance time of 24 rats in two groups on a rotating cylinder.

Usage

rotarod

Format

A data frame with 24 observations on 2 variables.

time

endurance time (seconds).

group

a factor with levels "control" and "treatment".

Details

The rats were randomly assigned to receive a fixed oral dose of a centrally acting muscle relaxant ("treatment") or a saline solvent ("control"). The animals were placed on a rotating cylinder and the endurance time of each rat, i.e., the length of time each rat remained on the cylinder, was measured up to a maximum of 300 seconds.

This dataset is the basis of a comparison of 11 different software implementations of the Wilcoxon-Mann-Whitney test presented in Bergmann, Ludbrook and Spooren (2000).

Note

The empirical variance in the control group is 0 and the group medians are identical. The exact conditional p-values are 0.0373 (two-sided) and 0.0186 (one-sided). The asymptotic two-sided p-value (corrected for ties) is 0.0147.

Source

Bergmann, R., Ludbrook, J. and Spooren, W. P. J. M. (2000). Different outcomes of the Wilcoxon-Mann-Whitney test from different statistics packages. The American Statistician 54(1), 72–77. doi: 10.1080/00031305.2000.10474513

Examples

## One-sided exact Wilcoxon-Mann-Whitney test (p = 0.0186)
wilcox_test(time ~ group, data = rotarod, distribution = "exact",
            alternative = "greater")

## Two-sided exact Wilcoxon-Mann-Whitney test (p = 0.0373)
wilcox_test(time ~ group, data = rotarod, distribution = "exact")

## Two-sided asymptotic Wilcoxon-Mann-Whitney test (p = 0.0147)
wilcox_test(time ~ group, data = rotarod)

[Package coin version 1.4-1 Index]