statistic-methods {coin}	R Documentation

Extraction of the Test Statistic and Linear Statistic

Description

Methods for extraction of the test statistic and linear statistic.

Usage

## S4 method for signature 'IndependenceLinearStatistic'
statistic(object, type = c("test", "linear", "standardized"), ...)
## S4 method for signature 'IndependenceTest'
statistic(object, type = c("test", "linear", "standardized"), ...)
## S4 method for signature 'IndependenceTestStatistic'
statistic(object, type = c("test", "linear", "standardized"), ...)

Arguments

object
an object from which the test statistic or linear statistic can be extracted.

type
a character, the type of statistic: either "test" (default) for the test statistic, "linear" for the unstandardized linear statistic or "standardized" for the standardized linear statistic.

...
further arguments (currently ignored).

Details

The method statistic extracts the test statistic or the, possibly multivariate, linear statistic in its unstandardized or standardized form.

The test statistic (type = "test") is returned by default. The unstandardized or standardized linear statistic is obtained by setting type to "linear" or "standardized" respectively.

Value

The test statistic or the unstandardized or standardized linear statistic extracted from object. A numeric vector or matrix.

Examples

## Example data
dta <- data.frame(
    y = gl(4, 5),
    x = gl(5, 4)
)

## Asymptotic Cochran-Mantel-Haenszel Test
ct <- cmh_test(y ~ x, data = dta)

## Test statistic
statistic(ct)

## The unstandardized linear statistic...
statistic(ct, type = "linear")

## ...is identical to the contingency table
xtabs(~ x + y, data = dta)

## Illustrating departures from the null hypothesis of independence using the
## standardized linear statistic
statistic(ct, type = "standardized")

[Package coin version 1.1-2 Index]