vision {coin}	R Documentation

Unaided Distance Vision

Description

Assessment of unaided distance vision of women in Britain.

Usage

vision

Format

A contingency table with 7477 observations on 2 variables.

Right.Eye

a factor with levels "Highest Grade", "Second Grade", "Third Grade" and "Lowest Grade".

Left.Eye

a factor with levels "Highest Grade", "Second Grade", "Third Grade" and "Lowest Grade".

Details

Paired ordered categorical data from case-records of eye-testing of 7477 women aged 30–39 years employed by Royal Ordnance Factories in Britain during 1943–46, as given by Stuart (1953).

Stuart (1955) used this data set to illustrate a test of marginal homogeneity.

Source

Stuart, A. (1953). The estimation and comparison of strengths of association in contingency tables. Biometrika 40(1/2), 105–110.

References

Stuart, A. (1955). A test for homogeneity of the marginal distribution in a two-way classification. Biometrika 42(3/4), 412–416.

Examples

## Asymptotic Stuart(-Maxwell) test (Q = 11.96)
diag(vision) <- 0 # speed-up
mh_test(vision)

[Package coin version 1.1-2 Index]