vision {coin}	R Documentation

Unaided Distance Vision

Description

Assessment of unaided distance vision of women in Britain.

Usage

vision

Format

A contingency table with 7477 observations on 2 variables.

Right.Eye

a factor with levels "Highest Grade", "Second Grade", "Third Grade" and "Lowest Grade".

Left.Eye

a factor with levels "Highest Grade", "Second Grade", "Third Grade" and "Lowest Grade".

Details

Paired ordered categorical data from case-records of eye-testing of 7477 women aged 30–39 years employed by Royal Ordnance Factories in Britain during 1943–46, as given by Stuart (1953).

This data set was used by Stuart (1955) to illustrate a test of marginal homogeneity. Winell and Lindbäck (2018) also used the data, demonstrating a score-independent test for ordered categorical data.

Source

Stuart, A. (1953). The estimation and comparison of strengths of association in contingency tables. Biometrika 40(1/2), 105–110. doi: 10.2307/2333101

References

Stuart, A. (1955). A test for homogeneity of the marginal distributions in a two-way classification. Biometrika 42(3/4), 412–416. doi: 10.1093/biomet/42.3-4.412

Winell, H. and Lindbäck, J. (2018). A general score-independent test for order-restricted inference. Statistics in Medicine 37(21), 3078–3090. doi: 10.1002/sim.7690

Examples

## Asymptotic Stuart test (Q = 11.96)
diag(vision) <- 0 # speed-up
mh_test(vision)

## Asymptotic score-independent test
## Winell and Lindbaeck (2018)
(st <- symmetry_test(vision,
                     ytrafo = function(data)
                         trafo(data, factor_trafo = function(y)
                             zheng_trafo(as.ordered(y)))))
ss <- statistic(st, type = "standardized")
idx <- which(abs(ss) == max(abs(ss)), arr.ind = TRUE)
ss[idx[1], idx[2], drop = FALSE]

[Package coin version 1.4-1 Index]